machine_learning_project/requirements.txt
biopython==1.76
h5py==2.10.0
Keras==2.3.1
numpy==1.18.2
pandas==1.0.2
scikit-learn==0.22.2.post1
scipy==1.4.1
tensorflow==1.14.0
